==> 3. Stem, HSS, Interdisciplinary network.Rmd <==
---
title: "3. STEM, HSS, Interdisciplinary network and Community Detection"
author: "Jieun Park"
date: "`r Sys.Date()`"
output: html_document
---

# MERGE

```{r}
# Interdisciplinary
Interdis <- rbind(ARTS, SOCI, COMP, ENGI, PHYS, MATHS)
write.csv(Interdis, "Interdis.csv")
```

# HSS and STEM

```{r}
Interdis <- read_csv("~/Desktop/master thesis⭐️/LaTeX/Disciplinary_graph_data/Interdis.csv")
# 1
network <- Interdis |> 
  filter(number_of_collaborator > 1) |> 
  group_by(title) |> 
  mutate(authors = list(authid)) |> 
  select(title, authors) |> 
  ungroup() |> 
  distinct(title, .keep_all = TRUE) |> 
  unnest(authors)


# 2
replace_null <- function(x) {
  ifelse(is.null(x), 0, 1)
}

# 3
# Pivot the data frame wider
wide_network <- pivot_wider(network, 
                             names_from = authors, 
                             values_from = authors, 
                             values_fn = replace_null)

# 4
wide_network <- ifelse(is.na(wide_network), 0, 1) 
wide_network <- data.frame(wide_network) |> select(-title)
wide_network_interdis <- as.matrix(wide_network)
wide_network_interdis_transpose <- t(wide_network_interdis)

dim(wide_network_interdis_transpose)
dim(wide_network_interdis)

# 5 
# Perform matrix multiplication - author projection making squared matrix
author_relationship_interdis <- wide_network_interdis_transpose %*% wide_network_interdis 

diag(author_relationship_interdis) <- 0
dim(author_relationship_interdis)

inter_matrix_zero_one = ifelse(author_relationship_interdis == 0, 0, 1)
dim(inter_matrix_zero_one)

# 6
# save the data
write.csv(inter_matrix_zero_one, "inter_matrix_zero_one.csv", row.names = FALSE)

```

# Giant Component

```{r}
inter_matrix_zero_one_data <- as.matrix(read.csv("Disciplinary_graph_data/inter_matrix_zero_one.csv"))
dim(inter_matrix_zero_one_data)
# inter_matrix_zero_one_data <- as.matrix(inter_matrix_zero_one_data)

g <- graph_from_adjacency_matrix(inter_matrix_zero_one_data)
```



```{r}
# Edges
edges = as_data_frame(g)

# Remove X from data
edges = data.frame(apply(edges, 2, function(x) gsub("X", "", x)))

edges <- edges |> filter(from != "3.5561e.10")
edges <- edges |> filter(to != "3.5561e.10")

edges

# Nodes
nodes <- Interdis |> select(authid, discipline)

nodes <- nodes |> mutate(discipline = ifelse(discipline %in% c("ARTS", "SOCI"), "HSS",
                             ifelse(discipline %in% c("PHYS", "COMP", "ENGI", "MATHS"), "STEM",
                                    discipline)))

nodes <- rbind(nodes[2, ], nodes[-2, ])
nodes$authid <- as.character(nodes$authid)

nodes <- as.data.frame(nodes)
  
# write.csv(nodes, "nodes.csv")
# write.csv(edges, "edges.csv")
```

```{r}
# who had only one disciplinary orientation
nodes_one <- nodes %>%
  group_by(authid) %>%
  count(discipline) %>%
  filter(length(authid) == 1) %>%
  ungroup()  |> select(authid, discipline)
  
# who had two disciplinary orientation
nodes_selected <-  nodes %>%
  group_by(authid) %>%
  count(discipline) %>%
  filter(length(authid) == 2) %>%
  ungroup() %>%
  group_by(authid) %>%
  filter(n == max(n)) |> 
  distinct(authid, .keep_all = TRUE) |> 
  select(authid, discipline)

nodes <- rbind(nodes_one, nodes_selected)
nodes
```


```{r}
# filter 2024
nodes_2024 <- nodes |> group_by(date, discipline, title) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2024)

# filter 2023
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2023)

# filter 2022
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2022)

# filter 2021
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2021)

# filter 2020
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2020)

# filter 2019
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2019)

# filter 2018
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2018)

# filter 2017
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2017)

# filter 2016
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2016)

# filter 2015
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2015)

# filter 2014
nodes_2024 <- nodes |> group_by(date, discipline) |> distinct(authid, .keep_all = TRUE) |> 
  filter(date == 2014)

```



```{r}
graph <- graph_from_data_frame(edges, vertices = nodes, directed=FALSE)

graph

write_graph(graph, "graph.graphml", format = "gml")
```

```{r}
components <- components(graph)
head(components$csize,10)

big_component <- igraph::decompose(graph)[[1]]

# save the graph
write_graph(big_component, "big_component.graphml", format='gml')
```

# Interdisciplinary

```{r}
# Load graph from Pajek
big_component_data <- read_graph("Disciplinary_graph_data/big_component.graphml", format = "gml")

# simplify the graph

gRC <- simplify(big_component_data)

diameter(gRC)
mean_distance(gRC)
transitivity(gRC)

vcount(gRC)

plot(gRC,
     vertex.size = 3,
     vertex.label = NA,
     edge.color = "black",
     edge.width = 0.8)


```

### Betweeness centrality

```{r}
betweenness_centrality <- betweenness(gRC)

# Find the maximum betweenness centrality value
max_betweenness <- max(betweenness_centrality)

# Find the node(s) with the maximum betweenness centrality
nodes_with_max_betweenness <- which(betweenness_centrality == max_betweenness)

# Check the names and values
node_ids <- names(nodes_with_max_betweenness)

# Print the node(s) with the highest betweenness centrality
print(paste("Node(s) with highest betweenness centrality:", node_ids))
print(paste("Highest betweenness centrality value:", round(max_betweenness, 0)))

graph_df <- get.data.frame(gRC, what="both")
graph_df$edges |> filter(from == 56249634600)

png("the node that has the highest betweenness centrality in the interdisciplinary network.png")

vertex.size <- ifelse(V(gRC)$name == 56249634600, 7, 2)
vertex.color <- ifelse(V(gRC)$name == 56249634600, "red", "black")

plot(gRC, 
     vertex.color = vertex.color,
     vertex.label = "", 
     vertex.size = vertex.size, 
     layout = layout_nicely)


dev.off()
```

### Closeness centrality

```{r}
closeness_centrality <- closeness(gRC)

# Find the maximum betweenness centrality value
max_closeness <- max(closeness_centrality)

# Find the node(s) with the maximum betweenness centrality
nodes_with_max_closeness <- which(closeness_centrality == max_closeness)

# Check the names and values
node_ids <- names(nodes_with_max_closeness)

# Print the node(s) with the highest betweenness centrality
print(paste("Node(s) with highest closeness centrality:", node_ids))
print(paste("Highest closeness centrality value:", round(max_closeness, 5)))

max_closeness <- 2e-05
formatted_closeness <- format(max_closeness, scientific = FALSE)
print(paste("Highest closeness centrality value:", formatted_closeness))

graph_df <- get.data.frame(gRC, what="both")
graph_df$edges |> filter(from == 6602128193)


png("the node that has the highest closeness centrality in the interdisciplinary network.png")

vertex.size <- ifelse(V(gRC)$name == 6602128193, 7, 2)
vertex.color <- ifelse(V(gRC)$name == 6602128193, "red", "black")

plot(gRC, 
     vertex.color = vertex.color,
     vertex.label = "", 
     vertex.size = vertex.size, 
     layout = layout_nicely)

dev.off()
```

### degree centrality

```{r}
# Calculate degree centrality
degree_centrality <- degree(gRC, mode = "all")

# Print degree centrality
print(degree_centrality)

# Get the node with the highest degree centrality
max_degree_node <- which.max(degree_centrality)
max_degree_value <- max(degree_centrality)

cat("Node with the highest degree centrality:", V(gRC)$name[max_degree_node], "\n")
cat("Degree centrality value:", max_degree_value, "\n")
```

### eigenvector_centrality

```{r}
eigenvector_centrality <- eigen_centrality(gRC)$vector
max_eigenvecto_centrality <- max(eigenvector_centrality)

cat("Node with the highest degree centrality:", V(gRC)$name[max_eigenvecto_centrality], "\n")
cat("Degree centrality value:", max_eigenvecto_centrality, "\n")
```



### Matching

```{r}
ARTS |> select(authid, authname) |> filter(authid == 6602128193)
SOCI |> select(authid, authname) |> filter(authid == 6602128193)

COMP |> select(authid, authname) |> filter(authid == 6602128193)
ENGI |> select(authid, authname) |> filter(authid == 6602128193)
PHYS |> select(authid, authname) |> filter(authid == 6602128193)
MATHS |> select(authid, authname) |> filter(authid == 6602128193)

```



### Distance

# Subgraph - HSS

```{r}
subgraph_HSS <- V(gRC)[V(gRC)$discipline == "HSS"]
g_sub_HSS <- induced_subgraph(gRC, subgraph_HSS)
HSS <- components(g_sub_HSS)
head(HSS$csize,10)

big_index_hss <- which(HSS$csize == max(HSS$csize))
big_index_nodes_hss <- which(HSS$membership == big_index_hss)

HSS_giant_component <- induced_subgraph(g_sub_HSS, big_index_nodes_hss)
HSS_giant_component <- simplify(HSS_giant_component)

vcount(HSS_giant_component)
round(mean_distance(HSS_giant_component, directed = FALSE, weights = NULL), 2)
diameter(HSS_giant_component, directed = FALSE, weights = NULL)
transitivity(HSS_giant_component)
```

### HSS centrality

```{r}

betweenness_centrality <- betweenness(HSS_giant_component)

# Find the maximum betweenness centrality value
max_betweenness <- max(betweenness_centrality)

# Find the node(s) with the maximum betweenness centrality
nodes_with_max_betweenness <- which(betweenness_centrality == max_betweenness)

# Check the names and values
node_ids <- names(nodes_with_max_betweenness)

# Print the node(s) with the highest betweenness centrality
print(paste("Node(s) with highest betweenness centrality:", node_ids))
print(paste("Highest betweenness centrality value:", round(max_betweenness, 0)))

graph_df <- get.data.frame(HSS_giant_component, what="both")
graph_df$edges |> filter(from == 16432831800)



plot(HSS_giant_component,
     vertex.label = "",
     vertex.size = 3)

vertex.size <- ifelse(V(HSS_giant_component)$name == 16432831800, 6, 2)

png("the node that has the highest betweenness centrality in the HSS network.png")

plot(HSS_giant_component, 
     vertex.color = ifelse(V(HSS_giant_component)$name == 16432831800, "red", "black"),
     vertex.label = "", vertex.size = vertex.size, layout = layout_nicely)

dev.off()
```

### Closeness centrality

```{r}
closeness_centrality <- closeness(HSS_giant_component)

# Find the maximum betweenness centrality value
max_closeness <- max(closeness_centrality)

# Find the node(s) with the maximum betweenness centrality
nodes_with_max_closeness <- which(closeness_centrality == max_closeness)

# Check the names and values
node_ids <- names(nodes_with_max_closeness)

# Print the node(s) with the highest betweenness centrality
print(paste("Node(s) with highest closeness centrality:", node_ids))
print(paste("Highest closeness centrality value:", round(max_closeness, 5)))

graph_df <- get.data.frame(gRC, what="both")
graph_df$edges |> filter(from == 55330575700)

png("the node that has the highest closeness centrality in the HSS network.png")

vertex.size <- ifelse(V(HSS_giant_component)$name == 55330575700, 6, 2)
vertex.color <- ifelse(V(HSS_giant_component)$name == 55330575700, "red", "black")

plot(HSS_giant_component, 
     vertex.color = vertex.color,
     vertex.label = "", 
     vertex.size = vertex.size, 
     layout = layout_nicely)

dev.off()
```


### degree centrality

```{r}
# Calculate degree centrality
degree_centrality <- degree(HSS_giant_component, mode = "all")

# Print degree centrality
print(degree_centrality)

# Get the node with the highest degree centrality
max_degree_node <- which.max(degree_centrality)
max_degree_value <- max(degree_centrality)

cat("Node with the highest degree centrality:", V(HSS_giant_component)$name[max_degree_node], "\n")
cat("Degree centrality value:", max_degree_value, "\n")
```

### eigenvector_centrality

```{r}
eigenvector_centrality <- eigen_centrality(HSS_giant_component)$vector
max_eigenvecto_centrality <- max(eigenvector_centrality)

cat("Node with the highest degree centrality:", V(HSS_giant_component)$name[max_eigenvecto_centrality], "\n")
cat("Degree centrality value:", max_eigenvecto_centrality, "\n")
```

### Searching the researcher's name

```{r}
ARTS |> select(authid, authname) |> filter(authid == 55330575700)
SOCI |> select(authid, authname) |> filter(authid == 55330575700)
```


# STEM subgraph

```{r}
subgraph_STEM <- V(gRC)[V(gRC)$discipline == "STEM"]
g_sub_STEM <- induced_subgraph(gRC, subgraph_STEM)
STEM <- components(g_sub_STEM)
head(STEM$csize,10)

big_index_STEM <- which(STEM$csize == max(STEM$csize))
big_index_nodes_STEM <- which(STEM$membership == big_index_STEM)

STEM_giant_component <- induced_subgraph(g_sub_STEM, big_index_nodes_STEM)

mean_distance(STEM_giant_component, directed = FALSE, weights = NULL)
diameter(STEM_giant_component, directed = FALSE, weights = NULL)
transitivity(STEM_giant_component)

plot(STEM_giant_component,
     vertex.label = "",
     vertex.size = 3)
```


```{r}
diameter(V(gRC)[V(gRC)$discipline == "HSS"])
transitivity(gRC)
assortativity_degree(gRC)
```

### betweenness centrality

```{r}
betweenness_centrality <- betweenness(STEM_giant_component)

# Find the maximum betweenness centrality value
max_betweenness <- max(betweenness_centrality)

# Find the node(s) with the maximum betweenness centrality
nodes_with_max_betweenness <- which(betweenness_centrality == max_betweenness)

# Check the names and values
node_ids <- names(nodes_with_max_betweenness)

# Print the node(s) with the highest betweenness centrality
print(paste("Node(s) with highest betweenness centrality:", node_ids))
print(paste("Highest betweenness centrality value:", round(max_betweenness, 0)))

graph_df <- get.data.frame(STEM_giant_component, what="both")
graph_df$edges |> filter(from == 36661053300)

png("the node that has the highest betweenness centrality in the STEM network.png")

vertex.size <- ifelse(V(STEM_giant_component)$name == 36661053300, 7, 2)
vertex.color <- ifelse(V(STEM_giant_component)$name == 36661053300, "red", "black")

plot(STEM_giant_component, 
     vertex.color = vertex.color,
     vertex.label = "", 
     vertex.size = vertex.size, 
     layout = layout_nicely)

dev.off()
```

### Closeness centrality

```{r}
closeness_centrality <- closeness(STEM_giant_component)

# Find the maximum betweenness centrality value
max_closeness <- max(closeness_centrality)

# Find the node(s) with the maximum betweenness centrality
nodes_with_max_closeness <- which(closeness_centrality == max_closeness)

# Check the names and values
node_ids <- names(nodes_with_max_closeness)

# Print the node(s) with the highest betweenness centrality
print(paste("Node(s) with highest closeness centrality:", node_ids))
print(paste("Highest closeness centrality value:", round(max_closeness, 5)))

max_closeness <- 2e-05
formatted_closeness <- format(max_closeness, scientific = FALSE)
print(paste("Highest closeness centrality value:", formatted_closeness))

graph_df <- get.data.frame(STEM_giant_component, what="both")
graph_df$edges |> filter(from == 6602128193)

png("the node that has the highest closeness centrality in the STEM network.png")

vertex.size <- ifelse(V(STEM_giant_component)$name == 6602128193, 7, 2)
vertex.color <- ifelse(V(STEM_giant_component)$name == 6602128193, "red", "black")

plot(STEM_giant_component, 
     vertex.color = vertex.color,
     vertex.label = "", 
     vertex.size = vertex.size, 
     layout = layout_nicely)

dev.off()
```

### degree centrality

```{r}
# Calculate degree centrality
degree_centrality <- degree(STEM_giant_component, mode = "all")

# Print degree centrality
print(degree_centrality)

# Get the node with the highest degree centrality
max_degree_node <- which.max(degree_centrality)
max_degree_value <- max(degree_centrality)

cat("Node with the highest degree centrality:", V(STEM_giant_component)$name[max_degree_node], "\n")
cat("Degree centrality value:", max_degree_value, "\n")
```

### eigenvector_centrality

```{r}
eigenvector_centrality <- eigen_centrality(STEM_giant_component)$vector
max_eigenvecto_centrality <- max(eigenvector_centrality)

cat("Node with the highest degree centrality:", V(STEM_giant_component)$name[max_eigenvecto_centrality], "\n")
cat("Degree centrality value:", max_eigenvecto_centrality, "\n")
```

```{r}
COMP |> select(authid, authname) |> filter(authid == 6602128193)
ENGI |> select(authid, authname) |> filter(authid == 6602128193)
PHYS |> select(authid, authname) |> filter(authid == 6602128193)
MATHS |> select(authid, authname) |> filter(authid == 6602128193)
```



# Community

```{r}
V(gRC)$color[V(gRC)$discipline == "HSS"] <- "red"
V(gRC)$color[V(gRC)$discipline == "STEM"] <- "blue"

# HSS
V(gRC)[V(gRC)$discipline == "HSS"] # 1177
# STEM
V(gRC)[V(gRC)$discipline == "STEM"] # 11464

length(V(gRC)[V(gRC)$discipline == "HSS"])
length(V(gRC)[V(gRC)$discipline == "STEM"])

layout <- layout_nicely(gRC)

png("plot_HSS_STEM_community.png")

par(mar=c(0, 0, 0, 0))
plot(gRC, 
     vertex.label = "",
     vertex.size = 3,
     layout = layout)

dev.off()


```


```{r}
community = cluster_louvain(gRC, resolution = 0.01)
head(membership(community))
modularity(community)
length(community)

V(gRC)$color[V(gRC)$discipline == "HSS"] <- "red"
V(gRC)$color[V(gRC)$discipline == "STEM"] <- "blue"

layout <- layout_nicely(gRC)

png("plot_interdisciplinary_network_community.png")

par(mar=c(0, 0, 0, 0))
plot(community, gRC, 
     vertex.label = "",
     vertex.size = 3,
     layout = layout) 

dev.off()

```


### Measuring Community Sizes for Each Clusters

```{r}
# Load necessary library
library(igraph)

# Assuming `gRC` is your graph object
# Perform community detection using the Louvain method

# Get membership of the first few nodes
head(membership(community))

# Calculate modularity
modularity(community)

# Get the number of communities
length(community)

# Get the sizes of each community
community_sizes <- sizes(community)

# Convert the community sizes to a data frame
community_sizes_df <- data.frame(
  CommunityID = seq_along(community_sizes),
  NodeCount = as.vector(community_sizes)
)

# Print the data frame
print(community_sizes_df)

```

### centrality - community

```{r}
degree_centrality <- degree(gRC)
betweenness_centrality <- betweenness(gRC)
closeness_centrality <- closeness(gRC)
eigenvector_centrality <- eigen_centrality(gRC)$vector

# Create a data frame with all centrality measures and community membership
centrality_df <- data.frame(
  Node = V(gRC)$name,
  Community = membership(community),
  Degree = degree_centrality,
  Betweenness = betweenness_centrality,
  Closeness = closeness_centrality,
  Eigenvector = eigenvector_centrality
)

# Aggregate centrality measures by community
numeric_centrality_df <- centrality_df[, -1] # Exclude Node column

# Aggregate centrality measures by community
community_centrality <- aggregate(. ~ Community, data = numeric_centrality_df, mean)


# Print the aggregated centrality measures for each community
print(community_centrality)

community_centrality_max <- aggregate(. ~ Community, data = numeric_centrality_df, max)

# Print the aggregated maximum centrality measures for each community
print(community_centrality_max)
print(round(community_centrality$Betweenness,2))


```

### How many STEM nodes and HSS nodes from each communities?

```{r}
node_data <- data.frame(
  Node = V(gRC)$name,
  Community = membership(community),
  Discipline = V(gRC)$discipline
)

# Count the number of STEM and HSS nodes in each community
community_discipline_count <- node_data %>%
  group_by(Community, Discipline) %>%
  summarise(NodeCount = n(), .groups = 'drop')

# Print the result
print(community_discipline_count)
```
